logs user to database on bot creation :)
