+++ src/core/ai/langchain/langchain.module.ts
@@ -4,11 +4,10 @@ import { GcpModule } from '../../gcp/gcp.module';
 import { LangchainService } from './langchain.service';
 import { OpenaiModule } from '../openai/openai.module';
 import { DatabaseModule } from '../../database/database.module';
-import { LangchainChainService } from './langchain-chain.service';
 
 @Module({
   imports: [GcpModule, OpenaiModule, DatabaseModule],
-  providers: [LangchainService, LangchainChainService, Logger],
-  exports: [LangchainService, LangchainChainService],
+  providers: [LangchainService, Logger],
+  exports: [LangchainService],
 })
 export class LangchainModule {}
