+++ src/core/bot/bot.dto.ts
