+++ src/interfaces/discord/services/discord-thread.service.ts
@@ -8,9 +8,7 @@ import {
   TextChannel,
   ThreadChannel,
 } from 'discord.js';
-import { ChatbotService } from '../../../core/chatbot/chatbot.service';
-import { InjectModel } from '@nestjs/mongoose';
-import { User } from '../../../core/database/schema/users/user.schema';
+import { BotService } from '../../../core/bot/bot.service';
 import { UsersService } from '../../../core/database/schema/users/users.service';
 
 @Injectable()
@@ -18,7 +16,7 @@ export class DiscordThreadService implements OnModuleDestroy {
   constructor(
     private readonly _usersService: UsersService,
     private readonly _logger: Logger,
-    private readonly _chatbotService: ChatbotService,
+    private readonly _botService: BotService,
     private readonly _client: Client,
   ) {}
 
@@ -52,8 +50,8 @@ export class DiscordThreadService implements OnModuleDestroy {
       discordId: interaction.user.id,
     });
 
-    // await this._chatbotService.createChatbot(thread.id);
-    await this._chatbotService.createChatbot(thread.id);
+    // await this._botService.createChatbot(thread.id);
+    await this._botService.createBot(thread.id);
 
     this._beginWatchingIncomingMessages(interaction, channel, thread);
     await this._sendInitialReply(interaction, channel, thread, text);
@@ -104,13 +102,9 @@ export class DiscordThreadService implements OnModuleDestroy {
     inputText: string,
     message: Message<boolean>,
   ) {
-    const tokenStream = this._chatbotService.streamResponse(
-      thread.id,
-      inputText,
-      {
-        topic: channel.topic,
-      },
-    );
+    const tokenStream = this._botService.streamResponse(thread.id, inputText, {
+      topic: channel.topic,
+    });
     thread.sendTyping();
 
     let initialReply: Message<boolean> = undefined;
