+++ src/interfaces/slack/slack.module.ts
@@ -3,11 +3,11 @@ import { SlackService } from './slack.service';
 import { SlackCommandMethodDiscovery } from './decorators/discovery';
 import { DiscoveryModule } from '@golevelup/nestjs-discovery';
 import { SlackBoltModule } from './bolt/slack-bolt.module';
-import { ChatbotModule } from '../../core/chatbot/chatbot.module';
 import { GcpModule } from '../../core/gcp/gcp.module';
+import { BotModule } from '../../core/bot/bot.module';
 
 @Module({
-  imports: [SlackBoltModule, DiscoveryModule, ChatbotModule, GcpModule],
+  imports: [SlackBoltModule, DiscoveryModule, BotModule, GcpModule],
   providers: [SlackService, SlackCommandMethodDiscovery],
 })
 export class SlackModule implements OnModuleInit {
